improved achievement manager
Inside the update progression status function I've added an extra else if statement that checks if the collectable type is equal to an achievement type. If this is the case it will loop over all the achievements that have been unlocked and increase the int amount.

TO DO:
Fix issue where the total amount of achievements remaining to unlock the all achievements achievement stays at 1