Visual progression bug fixed solved
Issue where the progression on the achievement object wasn't properly updating. This has been fix adjusting the checking method inside the update achievement object funciton. Based on if the collectable type is a collable or achievement as well as if this has a manual goal amount or not.

TO DO:
- Fix bug that cause the all achievements unlocked achievement to pop once you've only achievement one achievement