Update AchievementSystem.cs
Moved the update achievement status function call from the check achievement requirement status to the handle value update function inside the else statement. This is done so it can also visually update the current integer or float value on the achievement object itself if the progress is being shown on them.