Fix game data new game bug
Issue was that if you press new game, the game would reset it self but the game data file stays empty untill you've either collected an item or unlocked an achievement. This has been fixed by calling the save game event in the reset all achievements function after it is done resetting everything.