Update CollectableSOEditor.cs
I've made some adjustments to how the list is being displayed. The is collected with boolean now stays properly on screen, regardless of the with expanding of the list. This way I can strech out the inspector gui to see the ID's without it effecting the is collected part of it.