Fixed all collectables bug completion
When I added the over time collectable SO to the all collectables list for easy management, I noticed that it uses that collectable to add to the total of all collectables that shouldn't be happening. I fixed this by adding in a check if the collectable category isn't equal to none. That way it doesn't get counted towards the total amount in both Is All collectables requirement met, get all collectables progression, get unique category count and get collected amount per category